#14 panic fixed when a target data structure defines values on UNH
It's a hack/workaround. Good enough for now. It keeps the commit
changes minimal.

Problem was the order of decodeNodes in the decodeTree. The most
important decodeNode 'root' was added last. The fix adds 'root' in front
of other decodeNodes.

NOTE instead of using UNH as the root segment it should probably be
UNB.